several win32 warnings

+++ libs/InteractiveToolkit-AppKit/opengl/appkit-gl-base/include/appkit-gl-base/font/GLFont2VBO.h
@@ -17,7 +17,7 @@ namespace AppKit
         {
         public:
             GLVertexBufferObject vbo; ///< current VBO
-            size_t count;             ///< current vertex count in the vbo
+            int count;             ///< current vertex count in the vbo
 
             GLFont2VBO();
